test(helm-output): add script to test and update helm output
This script will render all current charts based on provided values files and compare the output with an 'expected output' yaml file. This is crucial to ensure that the rendered output is not changed unexpectedly. When the output is expected to change, you can update the outputs. Any changes in expected outputs should be accompanied by a chart version change.