feat(smilecdr): normalize resource names
There was previously no consistency to the naming of resources. This commit
adjusts them all to be `<releasename>-scdr-[<optionalidentifier>]`. The optional
identifier is used in cases where there is more than one of a particular
resource type.

Note: Changing this alters the 'expected output' files by changing resource names.
This is a breaking change, but will not pay it mind right now as this is still
pre-release.

* Update name in resource templates
* Update tests for new expected outputs