feat(smilecdr): remove extra labels from default values
There is no reason for these extra labels to be in the default values file
as they serve no technical purpose for the operation of the chart or the
application.
Previously, the labels.client was used for naming some resources, but
those parts of the template have already been changed to use the chart
release name (Which was typically set the same as labels.client so the end
result is the same)

Note: Changing this alters the 'expected output' files by removing labels.
It does not affect the selector labels.
This is a non-breaking change.

* Remove predefined labels from default values file
* Add section where users can set extra labels
* Update tests for new expected outputs