feat(smilecdr): add back default tag functionality
Previously the default image tag was defined in the default values file
and can be overriden by custom values. The original intent of the
boilerplate template created by `helm create` is to have this come from
the Chart.yaml file.

Note: Changing this alters the 'expected output' files as it noq quotes
the image tag.
This is a non-breaking change.

* Re-instate logic to pull default image tag from Chart.yaml
* Update tests for new expected outputs