feat(smilecdr): update Ingress definition logic
These new versions of the Helm Chart were only tested with the Nginx
Ingress Controller. To make it more compatible with other charts in
use, it needs to support the AWS ALB.

This change adds support for the AWS ALB using the AWS Load Balancer
Controller. It also updates the name used to select the Nginx Ingress
Controller as it was incorrectly set to `aws-lbc-nlb`. It was named this
way as it was being used with the AWS Load Balancer Controller, but this
is not a requirement for Nginx Ingress so it did not make sense.

Note: Changing this alters the 'expected output' files by quoting value
annotations. This is a non-breaking change resulting from the way the
various annotations are now concatenated.

* Add support for AWS ALB using `aws-lbc-alb`
* Add default set of annotations for AWS ALB
* Rename Nginx Ingress option from `aws-lbc-nlb` to `nginx-ingress`
* Update values file to reflect the above
* Update tests for new expected outputs

BREAKING CHANGE: Now uses `nginx-ingress` instead of
`aws-lbc-nlb` for specifying Nginx Ingress Controller