feat(smilecdr): add support for deploying postgres
Including the postgres manifests from the crunchydata examples repo.
This is a direct copy of the upstream charts, just with some variable
modifications so that we can include crunchypgo CRD values alongside
the SmileCDR values.

DO NOT USE THIS IN PRODUCTION
This is intended only for demonstration purposes and non-production
environments.