feat(smilecdr): add redeploy on config changes
This new feature is enabled by default, but can be disabled by
setting `autoDeploy` to false.

When configMaps are generated by Helm, their name includes a sha256 hash
of the data within. The deployments reference the config map by this
hashed name.

If a config map is updated, a new one is created and the deployment will
get updated as it points to the new one.

Due to having multiple configmaps, existing deployments are then
unaffected as they will not know about the new ones.

All expected output files are affected as this changes the default
behaviour

BREAKING CHANGES. This affects all existing deployments.