feat(smilecdr): add pod disruption budget
Using MaxUnavailable = 1 is a decent starting point for pdb
It allows draining of nodes, but also keeps the amount of simultaneous
disruption to a minimum.

In the scenario of there being 1 pod, the pod could get restarted during
a node draining event. This is reasonable, as if there is only 1 pod
running then there can be no expectation of resiliency or HA/FT.

If we chose MinAvailable = 1(for example), then we would never be able
to drain/service a K8s node with a pod on it if there was only 1 replica

Ideally, there should be 2+ pods running at all times as a minimum. Note
that if there are 2 pods and an eviction takes place, there will only
be a single pod, which means there is no longer any HA/FT. In this
scenario, it may be preferable to set MinAvailable to 2 or 3 if high
guarantees/SLAs are required.

This change updates ALL expected output files as it's adding a new
k8s manifest by default