feat(smilecdr): configure rolling deployments
This chance modifies the deployment strategy to use a rolling deployment
with a maxSurge of 1 and maxUnavailable of 0.

This means that an extra 1 pod will be created before an old pod is
destroyed. This is not the fastest deployment method, but it is a well
controlled one where it may be interrupted if the new pod does not come
up. In this case, the existing pods will remain unaffected.

This change updates ALL expected output files as it's adding a new
k8s manifest by default