feat(smilecdr): add support for strimzi kafka
With this change, you can now install a Kafka cluster alongside
the SmileCDR install.
SmileCDR is auto-configured to connect to the
kafka cluster.

DO NOT USE THIS IN PRODUCTION
This is intended only for demonstration purposes and non-production
environments.

All expected output files were updated due to removal of a comment
in the configMap
Added test for enabling Strimzi