feat(smilecdr): update JVM tuning params
This changes the behaviour of defining the java `-Xmx` settings.
It will now use the `request.memory` if defined, otherwise it will use
limits.memory.
`requests.memory` is deliberately left undefined in the default values
as K8s will automatically set it to be the same as `limits.memory`

This changes all expected output files as we no longer define
`requests.memory`