fix(smilecdr): allow quoted config entries
The properties file for SmileCDR does not allow quoted values,
however, some vales (e.g. `*`) cannot be passed in via
yaml and must be quoted.

This changes the logic so that strings can be quoted in the
yaml, but end up unquoted in the cdr config file.

New tests were added to the *modules* tests for this.