feat(smilecdr): change db secret config
This change makes it so that the external DB config is defined in the
same way as the image pull secrets.
i.e. you now specify `credentials.type` to be `sscsi` or `extsecret`.
If using `sscsi` you can also set the Secrets Store CSI Driver provider
with `credentials.provider`. e.g. `aws`.

This change does not change any expected outputs, but does require a
change to the values file.

BREAKING CHANGE: Values file needs to be updated if using sscsi for DB
secrets