feat(smilecdr): copy files from external location
This new feature allows copying files into the pod before the main Smile
CDR container starts up.
It uses multiple init containers as follows:

* `init-sync-classes`
This container is enabled by default if copying files to the `classes` dir.
It can be disabled using
```copyFiles.classes.disableSyncDefaults: true```.
This will copy the default files from the `classes` directory from the
Smile CDR base image to a `classes` shared volume that is local to
the pod

* `init-pull-classes-*`
This will copy files from the specified location to the `classes` shared
volume that is local to the pod. Any files copied will be available to
Smile CDR when it starts up.

* `init-pull-customerlib-*`
This will copy files from the specified location to the `customerlib`
shared volume that is local to the pod. Any files copied will be available
to Smile CDR when it starts up.

Currently the `init-pull` containers only support copying files from Amazon
S3 and websites using curl. When pulling from Amazon S3, an appropriate IAM
role must be configured via the ServiceAccount and IRSA.

All expected outputs changed as there is now an empty "initContainers"
list in the deployment pod spec even if this feature is not enabled.
This is not a breaking change.