feat(smilecdr): add support for Alpine3 base image
This change is in preparation for the next release that will run rootless
containers and use the Alpine3 base image.

The `startupProbe` used binaries in `/usr/bin/` which does not exist in
Alpine3 based images.
Using `/bin/` works in Alpine3 and in AL2 images, so no logic is needed
to determine which image is in use.

This change affects all expected outputs, but does not affect
functionality in manual testing. The pod still gets marked as ready
once the log output has the startup completion message.