fix(smilecdr): remove short-circuit dependency
In versions of Helm prior to 3.10.0, conditional expressions did not use
short-circuit behaviour. This means that in some cases there could be
a nil pointer exception when using these versions of Helm.

This change removed the dependency on short-circuit evaluation that is
provided in Helm 3.10.0 and higher. This change does not affect any
tests as they use Helm `^3.10.1` (`>= 3.10.1, < 4.0.0`) which do not
show this behaviour.