fix(smilecdr): use camelCase for `useDefaultModules`
The case of `usedefaultmodules` is inconsistent with other Helm and K8s
manifests in general and causes confusion as to which should be used.
The docs were also inconsistent in their usage of camelCase vs lower
case.
All tests are affected by this change

BREAKING CHANGES! Any reference to `usedefaultmodules` needs to be
updated to `useDefaultModules`