refactor(smilecdr): fix confusing case for nodeId
In K8s, camelCase is preferred. In Helm, UpperCamelCase is used for
reserved values. As such, having the default nodeId key use `Masterdev`
is inconsistent.

Rather than relying on the keyname for the nodeId, support has been
added for a 'name' field so that a `nodeId` starting with an upper case
character can be used if required.

A new test was created for this. No existing tests have been affected
and this is not a 'published' feature yet.