refactor(smilecdr): fix `secretArn` case
The use of `secretarn` and `secretArn` was inconsistent, leading to
confusion on when to use which style. All occurrences now use camelCase
in a way that acronyms should be treated as words and capitalized.
Code, tests and documentation have been updated to provide consistency.

BREAKING CHANGES! Any values files that use `secretarn` will need to be
updated.