feat(smilecdr): allow extra env vars and volumes
Allow custom configuration of extra environment variables, volumes and
volume mounts to be passed in. These are then merged with the auto-
generated ones.
Right now, these are specified at the global scope, but later it will be
possible to override them on a per-node basis.

Created tests for this feature.