feat(smilecdr): refactor image pull secrets
The mechanism for specifying image pull secrets was fairly broken as
mentioned in #78:

* Could only provide a single secret
* No option to not specify a secret
* YAML schema was cumbersome
* Used a list of secret definitions, but only used the first in the list

This change brings an overhaul to how image pull secrets are specified.
It has been done in a way to not break the existin mechanism, but it has
been marked as deprecated to allow time for administrators to update
to the newer mechanism without breaking existing environments.

You can now specify multiple image pull secrets. Each secret can be
defined in any of the supported methods (sscsi, k8sSecret or values).

Any that are defined with type `sscsi` will automatically be included in
the `SecretProviderClass` object.
Any that are defined with type `values` will automatically have `Secret`
objects created.

If there are multiple secrets defined, then a suffix is created to
prevent naming clashes.

BREAKING CHANGE: Deprecation warning - Values files must be updated to use
`image.imagePullSecrets` instead of `image.credentials`.