feat(smilecdr): improve warnings for chart errors
This change adds some checks for missing/incorrect `secretArn` entries
that could cause confusion when using newer versions of the chart with
older versions of the configuration.

As the chart is starting to be used more, these breaking changes need
to be surfaced more easily. As such, a new warning function has been
added to that it is easy to provide warnings and deprecation notices
to the administrator who is installing the Helm Chart.

This change does not affect any tests. The testing framework does not
currently support checking for error states or deprecation notices, so
no tests have been added for this feature.