fix(smilecdr): update initContainer configurations
For the `sync` init containers, the `cp` command was trying to sync the
properties of the mount point itself, which is read only.
Changed the `cp` command to avoid this.

The `copy` init containers were running as the container default user.
Instead, use the same `SecurityContext` for all containers so they all
run as the same user.
This change also improves the security stance as all init containers now
have a secure context.

Updated the appropriate tests.