feat(smilecdr): add HL7v2 support
This change adds support for the HL7 v2.* Listener Endpoint module that
requires special configuration.

In order to support this module we needed to add support for multiple
hostnames in the `Ingress` rules section. This was achieved by:

* Adding hostname to generated `service` objects
* Using default `context_path` if not provided by module
  This part was required, because specifying the `context_path` for the
  HL7 v2.* module would cause the module to throw errors during startup
  As a result, we cannot specify `context_path`.
* Created template to generate `host` entries for `Ingress` rules.
* Created tests for this feature.

Note that all tests were updated for this change due to altered ordering
of the entries in the `rules` objects. This had no change in semantics,
so does not affect the functionalily of the rendered manifests.