fix(smilecdr): use correct labels for Kafka admin
The Kafka Admin pod was using the same selector labels as the Smile CDR
pods. This caused the `Service` to send requests to the Kafka Admin pod,
which should not happen.
This change fixes this issue by setting a unique name for the Kafka Admin
pods.

Updated tests to reflect this change.
Also two kafka admin tests were misconfigured, so they were updated.