feat(smilecdr): add database properties mode
This change adds an experimental and unsupported feature that allows
you to use DATABASE mode for the properies source in Smile CDR.

This should only be used in cases where there is no way to test changes
using the values file or using the relatively safe `troubleshooting`
option.

This option is disabled by default and does not affect any existing
functionality.