fix(smilecdr): add preStop delay
Introduces a 30s delay before the CDR process gets terminated. This
allows time for load balancers to remove the pod from their targets,
preventing `HTTP 503` errors during scale-in or deployment rolling
events.
Total timeframe for the graceful shutdown (Default 30s) has also been
increased so that the application still has time to shut down gracefully
after the preStop delay has completed.

This changes the default settings for the deployment, so ALL tests are
affected by this change and have been updated accordingly.

Although this changes the default configuration of a pod, it only slows
the pod termination process so should not be considered a breaking
change.