fix(smilecdr): fix value for grace period
This error was introduced in 1.0.0-pre-65.
Although it was identified as a change in the default outputs, it was
not identified that the resulting k8s manifest was invalid.
This highlights a requirement to add K8s manifest validation to the
testing framework.

This change alters (fixes) all test output files.