fix(smilecdr): fix s3 copy with readonly rootfs
When the s3 functionality was added, the read only root fs was not yet
enforced. When this became enforced, it broke the aws-cli container
which caches temporary credentials in /.aws

Rather than making the aws-cli FS read-write, this change adds an
emtyDir mount to `/.aws`. Mounting this directory is the reccomendation
on the aws-cli dockerhub page.

The tests for s3 output have been updated.
In addition, the `customImage` test has been updated to not rely on the
s3 copy feature