fix(common): remove `chart.shortname` template
When using multiple charts in an umbrella chart, we cannot use
the `chart.shortname` template as the namespaces
clash.

This results in anything using this common template getting the
wrong value.
This manifests itself as any Image Pull Secrets all using the Directus
short name, for example.

* Removed `chart.shortname` template from any charts using it
* Added  a `shortName` entry in the values file for all charts
* Update references to `chart.shortname` to use `.Values.shortName`
instead

No tests were affected as this only shows up when using an umbrella
chart.