fix(smilecdr): allow hierarchical config
Previously, providing hierarchical module configurations resulted in
errors due to the way the configs are flattened for the java properties
file.
This change restructures the internal representation of the modules
configuration to be hierarchical. Any configurations that are provided
in flat form will be converted.

Configurations are then converted back to flat as needed during the
generation of the java properties file. The result of this is the
ability to define configurations in either style.

Flattened style is convenient for those migrating configurations from
existing Smile CDR properties files, while hierarchical eases integration with
external tools such as Terraform.

This is NOT a breaking change.