fix(smilecdr): allow specifying db name
When using the Crunchy PGO operator, this change allows you to specify
the db name.
By default, it will use the module name, or the user name if no module
name is supplied.

Now you can specify the db name with `dbName` and you can specify
the suffix with `dbSuffix`.

No breaking changes.