fix(smilecdr): add error checking for mapped files
Previously, if `mappedFiles` were defined, but no file was provided
using the Helm `--set-file` option, then the chart would deploy, but
the pod would never start as the ConfigMap would not be created.

This change will make the chart rendering fail with a warning so the
error can be caught early.