feat(smilecdr): add multi-node configuration
Multi-node configurations are here!

This feature allows you to configure Smile CDR in a 'multi-node'
configuration. There was a lot of changes to a lot of the Helm Chart
in order for this to function.

The documentation for this is still in progress, but the code is
currently working. This is still in an EXPERIMENTAL phase and
is not recommended for production usage.

This feature is currently `gated` by an option to use `oldResourceNames`
Part of the change when using this feature is that a lot of resource
names needed to be updated. As this causes a breaking change for
existing users of the chart, this compatibility mode was temporarily
added.

As the `oldResourceNames` toggle is currently set to `true`, this becomes
a non-breaking change. Existing users of the chart should not be affected.

All expected helm outputs have changed as the ordering of some items has
changed. There is no semantic differences in these outputs except for
the following:

* Database tests health endpoint changed. This was due to an incorrect
double-slash before the change, so this is a 'fix'
* Nodesettings tests had to be updated