refactor(pmp): update component generation
This change updates how the definitions for the multiple components
are defined. They are now more like the upcoming 'multi-node'
configurations in the Smile CDR chart.

This was initiated due to some weirdness in how the root context was
merged before. The result was that .Chart.Name was null as it was
being made available as .Chart.name instead which was confusing and
required special conditions to check which version was present.