refactor(smilecdr): use full context for nodes list
Before this change, some templates required a 'root' context
that contains `Release`, `Chart` and `Values` keys, amongst
other things. Whereas some other templates were written to
just use a passed in `Values` object.

This made it confusing when using a template as it was hard
to determine which uses which method.

In addition to this, there are some templates in upcoming
features that expected access to the root context, but only
had access to the node Values/Spec that was being passed in,
so they did not work.

With this change, the list of node definitions created by the
`smilecdr.nodes` helper template now returns a dict with
modified `root` context objects rather than just the .Values
for that CDR Node.

This means that any templates consuming that helper template
also needed to be modifuied.

No test Helm output files are affected by this change.