fix(smilecdr): multiple init-pull containers
When multiple file sources of the same type (i.e. `curl` or `s3`) are
defined, the initContainers now have unique names.

As this only takes effect when multiple sources are defined, it does
not affect the name when only a single init-pull container is created.

New tests have been added for this functionality and no
other tests have been affected.