feat(smilecdr): allow custom image repos
Previously, many of the supporting containers had hard coded image
repository settings. This makes it challenging to run in an environment
with restricted access to the public internet.

With this feature, a number of repository settings have changed:
* No longer using dockerhub for aws-cli and curl images.
  * curl now uses quay.io
  * aws-cli now uses public.ecr.aws
* No longer uses alpine for Smile CDR license copying.
  * Uses Smile CDR image instead
* Can now override repo for curl and s3 initContainers
* Can now override repo for Kafka Admin pod
* Can now override IAM Auth Jar provisioning for Kafka/MSK
  * Can use s3 file copy mechanism instead of default curl mechanism

* curl default image tag has been bumped to 8.1.2
* aws-cli default image tag has been bumped to 2.13.19
* IAM Auth Jar has been bumped to 1.1.9

This set of changes enables the functioning of this chart when running
on a cluster where downloading images & files from the internet has
been blocked.