fix(smilecdr): Remove duplicate env vars
When using Helm based auto-provisioning of thr OpenTelemetry
Java agent, the following env var gets defined twice:
`OTEL_RESOURCE_ATTRIBUTES_K8S_DEPLOYMENT_NAME`

This was not picked up on the intial release as this configuration
was not deployed to an environment. As a result, the tests also included
the variable twice and it was not picked up.

Tests have been updated.