fix(smilecdr): improve handling of issuer.url
The `issuer.url` config item is dealt with as a special case
as it's possible to provide the value `default` which will
auto-configure the url. However, this only works in the context of an
endpoint module and it does not allow overriding the value.

This change fixes this behaviour by asserting that the module is an
endpoint module if using 'default', and allowing for an
overridden value.

New test added for this scenario. No existing tests are affected.
Not a breaking change.