refactor(smilecdr): update probe generation
This change slightly refactors how the various pod probes are generated.

Startup Probe
This was previously hard-coded. Now it's configurable in the same way as
the readiness probe. This effects the generated YAML quoting, of the
startup command, but is semantically the same and is tested to still
work.

Readiness Probe
This was previously having the template helper generate text (YAML)
output, that was then converted with fromYaml and back to YAML again.
Directly generating the text was awkward to code with and inconsistent
with other areas of the template.

Liveness Probe
There was previously no support for liveness probes. This can now be
configured in the values file if required.

New tests added for the above.

ALL test outputs changed, but they are semantically identical.
Only quoting and probe ordering has changed.