docs(terraform): add initial Terraform module
This change adds the new Terraform module that can be used to provision
the required dependencies to use this Helm Chart in an EKS cluster.

This is currently an undocumented feature, as the Terraform module is still
under active initial development.

This will not bump the Helm Chart version or output and is not a breaking change.