repo(tf): move TF module to separate branch
When pulling TF modules from the main or pre-release branch, TF pulls the entire repo including the Helm Charts
and Docs.

This is inefficient and unnecessary. With this change, the terraform branch can be referenced and TF will only pull
the required TF module code.

This commit removes most of the non-tf related code. The pipeline will need to be updated to build a release process
for the TF code.
For now, any push to the TF module will update the live branch.