fix(smilecdr): update Strimzi schema
This was meant to be a simple fix to allow the
setting of the delete.topic.enable which was always
false by mistake. However, the Schema was confusing and and did not align with the Strimzi KafkaCluster CRD.

As a result, the Strimzi schema and some of the logic has been re-worked.