fix(smilecdr): auto-set s3 cp recursive option
THis change automatically uses the `--recursive` option for the `aws s3 cp` command if the source ends with a trailing slash.

This is due to the inability of the AWS CLI to use the recursive option if specifying an exact S3 object name.

Some Helm Output tests have been updated to ensure that the trailing slash is used where appropriate.
Some tests were updated as they were in error before this bug fix.

Breaking Change - If you currently specify a source path without a trailing slash, it will no longer try to recursively copy the files.