refactor(pmp): archive PMP helm charts
* Archived and removed all PMP related Helm charts and tests.

As these Helm Charts are not in use, nor in active development, they have been moved to the `archive-pmp` branch and removed from the main pre-release branch.

The inclusion of them causes extra delays while linting and building the main Smile CDR Helm Charts.

They can be pulled back in if development resumes on these charts.