refactor(smilecdr): update port config generation
The config for container ports was still using an older coding style which makes it challenging for the following future changes:
* Exposing prometheus agent port
* Adding arbitrary extra port definitions

Unit Tests:
This change uopdates ALL unit tests as it changes the order of the elements in the container port definitions.
There is no semantic difference in the change. This is not a breaking change.