feat(grafana): update dashboards to work with loki3
Loki 3 and the otlphttp log exporter gives us a different set of labels by default, so dashboard is updated to suit.

A big part of this is that the 'level' is no longer indexed by default. There are suggestions that you should not index this as you then get a separate `stream` for each level, which makes the chunk storage less efficient.

This changes the query logic when filtering by log level as it's now a filter rather than a selector.