fix(smilecdr): set java.io.tmpdir to correct location
Previously this picked up the default location of /tmp which is not writable due to the root filesystem being read only.

Any component in Smile CDR that tries to create temp files would fail due to this.

Adding an extra volume mount to this location would result in multiple TEMP directories which may cause future confusion.

Instead, we now set the JVM temp directory to /home/smile/smilecdr/tmp which is already mounted to an emptyDir volume.

This change alters ALL the helm check outputs, as it is a new default behaviour, but it is not a breaking change.