feat(smilecdr): allow ephemeral volume configuration
This change allows the size limits to be configurable for the following Smile CDR directories when they are mounted to 'emptyDir' volumes.

* /home/smile/smilecdr/tmp
* /home/smile/smilecdr/classes
* /home/smile/smilecdr/customerlib
* /home/smile/smilecdr/activemq-data

These are now configurable using:
```
volumeConfig:
  cdr:
    tmp:
      size: 10Gi
    classes:
      size: 10Gi
    customerlib:
      size: 10Gi
    log:
      size: 10Gi
```

Deprecation warning.
The `/home/smile/smilecdr/log` directory was already configurable using the `.Values.logsDirSize` option.
This option has been deprecated and a Helm Chart warning is issued if the option is still used.