fix(smilecdr): fix smileutil command
The new smileutil script included with this change is similar to the original
but with the following changes:

* Removed tput command (Only used for defining output column width)
* Use correct temp directory (/home/smile/smilecdr/tmp)
* Define `smile.basedir` java property (Resolves errors when running the command)
* Fix to still work if `JAVA_TOOLS_OPTIONS` has been set in this environment.

This change updates all outputs as it includes a custom `smileutil` script that functions correctly in
a Kubernetes pod. While technically this is a change to the deployment, it should not be considered to
be a 'breaking' change as there is no change in behaviour except that the smileutil command now works
as expected.