feat(smilecdr): update default context path for admin json module
All Helm Test outputs are updated by this change.

Breaking Change: If the default modules are being used, and if the context path for the admin json module has not been explicitly defined, then the context path for this module will change from `/json-admin` to `/admin_json` to align with the default naming of the other modules.
If you need it to remain as `/json-admin`, then set it using `modules.admin_json.config.context_path: json-admin`