fix(smilecdr): fix environment variable aggregation logic
Previously, the environment variable aggregation logic was performed using a list. This meant that it was more challenging to deal with duplicate scenarios.

This change updates the logic to use a dict. This way, if a duplicate is encountered, it will overwrite any existing entries rather than duplicating them.

This is a desired behaviour as the extraEnvVars comes later on in the aggregation logic, allowing one to explicitly override any existing auto-generated env vars.

This is with the exception of the JVMARGS env variable, which should not be manually set. This is because the JVMARGS are auto-configured to ensure that essential settings are in place (e.g. heap size and temp directory).

A warning has been added if the user has attempted to set JVMARGS within the extraEnvVars section of the values file.

In addition to the reported bug, the example provided in the docs [here](https://smilecdr-public.gitlab.io/smile-dh-helm-charts/latest/guide/smilecdr/envvars/#multi-node-configuration) also did not work. The global extraEnvVars would be completely overwritten if it was set in a per-node context.

This was fixed by passing the default global values into any helper functions so they can explicitly copy if required, and one of the multi-node helm tests was updated to capture this functionality.

All Helm Tests are updated with this change, but it is merely a change in ordering of the environment variables due to the updated internal representation. There is no semantic difference in the outputs.

Docs have also been updated to explain all of the above with more clarity.