ci(release): update release process
Prior to the v1.0.0 release, there was a single release channel called `devel`. Releases were auto-incremented using Semantic Release.
This was working well for the prerelease phase of this project, but since releasing v1.0.0, this needed some attention.
Initial attempt was to just create a `stable` channel and keep the `devel` channel for prereleases.
After 3 days to get this documented and working, fundamental problems were found with this method not being compatible with how the Semantic Release is designed to work.

The current version works like this:
* Release pipeline runs on `stable` branches, main|release-n.x
* Release pipeline runs on `prerelease` branches, pre-release, next-major, beta or alpha
* Helm Chart version is no longer updated in the repo. This has a number of benefits:
    * No need to run helm output update which takes ~8 mins
    * No need to update chart and write back to repo
    * Can now be published by Semantic Release (Used to be a separate pipeline)
* Jobs only run if files updated (This negates the requirement to use skip-helmcheck)
    * This speeds up pipelines for non-helm changes, as well as pipeline testing.
* Release channels are now parameterized in each job.
    * Actually simplifies the pipeline
    * No longer need separate jobs for stable and release channels