revert(sdh-common): remove common sub-chart
The sdh-common sub-chart has not been used or tested in a long time. It only muddies the releases by always being included.

Removing it for now, but may be re-introduced at a later time if required.

This change is not breaking as the sub-chart was not in active use anywhere.