fix(tf): pin s3 bucket dependency
Due to upstream changes that are not currently compatible with this terraform module, we need to pin the version of the `terraform-aws-modules/s3-bucket/aws` module.

This is because the newer version requires AWS provider >= 5.70.0

However this module is currently pinned to ~> 5.60.0 due to incompatible changes in the 5.61 version of said provider.

This mismatch causes the terraform init stage to fail.

This is a temporary fix that ultimately needs to be resolved by fixing the module limitations that cause >= 5.61 to fail.