feat(smilecdr): disable oldResourceNaming by default
Remove the default for oldResourceNaming.

Deployments will use the new resource naming unless manually overridden.

Breaking Change: This will change many resource names. There has been a deprecation warning for a long time now to warn users of this impending change. If someone needs to revert to the oldResourceNaming, they can still manually set it. The deprecation notice has been updated to indicate that it will be completely removed in an upcoming release.