feat(smilecdr): use AWS Advanced JDBC Driver configurations
As of Smile CDR `2025.05.R01`, the AWS Advanced JDBC Driver is used for IAM authentication
or when using Sectrets Manager. This requires a different method of configuring databases than
before.

This change uses the new CDR version checking functionality to determine which database config
mechanism should be used when using IAM auth or Secrets Manager.